remove MyDbRow
